Rename qubes.devices.BlockDevice to qubes.storage.BlockDevice
